Filter RSE name in RSEQuotaTable
Moved common `Filter` component into `Filter.tsx` within
`StreamedTables` folder.

Still need to work on responsive webdesign, since the textbox does not
take well to the resizing.
